refactor: コンポーネントのファイル名が page.tsx になっていたため、index.tsx に変更
